Fix compile issue
Currently the main branch fails to compile with the following error:

```
 --> /home/arjo/workspaces/rust/mapf/mapf/src/occupancy/mod.rs:159:14
    |
159 |         self.set_bit(corner.into(), status);
    |              ^^^^^^^ method not found in `&mut CornerStatus`
    |
    = help: items from traits can only be used if the trait is in scope
help: the following trait is implemented but not in scope; perhaps add a `use` for it:
    |
18  | use bitfield::BitMut;
    |
```

This tiny commit applies the above change. I will follow this up with a github ci action to check we don't break the main branch.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>